bot/requirements.txt
aiogram
Levenshtein
python-dotenv
sqlalchemy
psycopg2
